requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
python-multipart==0.0.9
